Trait jf_prf::PRF

pub trait PRF {
    type Input: Clone + CanonicalDeserialize;
    type Output: Clone + Debug + PartialEq + Eq + CanonicalSerialize;
    type Seed: Clone + Debug + Default + UniformRand + CanonicalSerialize + CanonicalDeserialize;
    type Error: Error;

    // Required method
    fn evaluate<S: Borrow<Self::Seed>, I: Borrow<Self::Input>>(
        seed: S,
        input: I
    ) -> Result<Self::Output, Self::Error>;

    // Provided method
    fn evaluate_with_rand_seed<R: RngCore + CryptoRng, T: Borrow<Self::Input>>(
        rng: &mut R,
        input: T
    ) -> Result<(Self::Seed, Self::Output), Self::Error> { ... }
}

Trait for Pseudo-random Functions

Required Associated Types

type Input: Clone + CanonicalDeserialize

Input to the PRF

type Output: Clone + Debug + PartialEq + Eq + CanonicalSerialize

Output of the PRF

type Seed: Clone + Debug + Default + UniformRand + CanonicalSerialize + CanonicalDeserialize

The random seed/key that index a specific function from the PRF ensembles

type Error: Error

Error type

Required Methods

fn evaluate<S: Borrow<Self::Seed>, I: Borrow<Self::Input>>( seed: S, input: I ) -> Result<Self::Output, Self::Error>

Compute PRF output with a user-provided randomly generated seed

Provided Methods

fn evaluate_with_rand_seed<R: RngCore + CryptoRng, T: Borrow<Self::Input>>( rng: &mut R, input: T ) -> Result<(Self::Seed, Self::Output), Self::Error>

same as Self::evaluate except that we generate a fresh random seed for the evaluation

Object Safety

This trait is not object safe.